# backend/api/openapi.yaml
openapi: 3.1.0

info:
  title: Stash API
  version: 0.1.0

paths:
  /users:
    post:
      summary: Create user
      operationId: createUser

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  maxLength: 72

      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                properties:
                  id:
                    type: string
                    format: uuid

        "409":
          description: User already exists

        "422":
          description: Invalid request

  /login:
    post:
      summary: Log in
      operationId: login

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string

      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"

        "401":
          description: Invalid credentials

  /refresh:
    post:
      summary: Refresh an access token
      operationId: refreshToken

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string

      responses:
        "200":
          description: Refresh successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"

        "401":
          description: Invalid refresh token

  /items/text:
    post:
      summary: Save text content
      operationId: createTextItem

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  minLength: 1

      responses:
        "202":
          description: Item accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemCreated"

        "401":
          description: Unauthorized

        "422":
          description: Invalid content

  /items/image:
    post:
      summary: Save image content
      operationId: createImageItem

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  maxLength: 10485760
                text:
                  description: >-
                    Optional caption stored on the same item (returned as its
                    `text`, and searchable). Blank is treated as absent.
                  type: string
            encoding:
              file:
                contentType: image/png, image/jpeg, image/gif, image/webp

      responses:
        "202":
          description: Item accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemCreated"

        "401":
          description: Unauthorized

        "422":
          description: Invalid content (empty file, unsupported image type, or file too large)

  /items/file:
    post:
      summary: Save a file
      description: >-
        Stores any file as-is, up to 50 MB. Analyzable formats (PDF,
        DOCX/XLSX/PPTX, ODT/ODS/ODP, RTF, EPUB, FB2, FB2.ZIP and the
        text/data formats) are created `pending` and analyzed asynchronously
        into a searchable description; everything else is created already
        `completed`. Recognized formats — documents (PDF,
        Office, OpenDocument, RTF, iWork), books (EPUB, FB2, FB2.ZIP,
        MOBI/AZW/AZW3, DjVu), text/data (TXT, MD, CSV, TSV, JSON, YAML, XML,
        HTML, LOG) and ZIP — keep their real content type, taken from the
        filename's extension and verified against the content. Any other
        file, or one whose content doesn't match its extension, is stored as
        `application/octet-stream`. The client's content type is ignored.
      operationId: createFileItem

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  maxLength: 52428800
                text:
                  description: >-
                    Optional caption stored on the same item (returned as its
                    `text`, and searchable). Blank is treated as absent.
                  type: string

      responses:
        "202":
          description: Item created (`pending` if it will be analyzed, otherwise `completed`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemCreated"

        "401":
          description: Unauthorized

        "422":
          description: Invalid content (empty file, or file larger than 50 MB)

  /items/{item_id}:
    delete:
      summary: Delete an item
      description: >-
        Permanently deletes the item, everything stored with it (text,
        caption, description, image file), and removes it from search.
      operationId: deleteItem

      security:
        - bearerAuth: []

      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "204":
          description: Item deleted

        "401":
          description: Unauthorized

        "404":
          description: No such item for the current user

        "422":
          description: Invalid item id

  /items:
    get:
      summary: List saved items
      operationId: listItems

      security:
        - bearerAuth: []

      parameters:
        - name: cursor
          in: query
          required: false
          description: Opaque pagination cursor from a previous response's `next_cursor`.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30

      responses:
        "200":
          description: The current user's items, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListItemsResponse"

        "401":
          description: Unauthorized

        "422":
          description: Invalid cursor

  /search:
    post:
      summary: Search saved content
      operationId: searchItems

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  minLength: 1
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 20

      responses:
        "200":
          description: >-
            Semantic search: the current user's items closest in meaning to
            `query` (by embedding cosine similarity, across languages),
            most similar first; clearly unrelated items are left out. Items
            whose content hasn't been embedded yet (a few seconds after
            saving/analysis) aren't found. Same item shape as `GET /items`.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ListedItem"

        "401":
          description: Unauthorized

        "422":
          description: Invalid request

        "503":
          description: Search is temporarily unavailable (the query couldn't be embedded)

components:
  schemas:
    TokenPair:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
        refresh_token:
          type: string

    ItemCreated:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed

    ListedFile:
      type: object
      required:
        - filename
        - content_type
        - size_bytes
      properties:
        filename:
          description: The original filename, as uploaded.
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer

    ListedItem:
      type: object
      required:
        - id
        - type
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - text
            - link
            - image
            - file
        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed
        created_at:
          type: string
          format: date-time
        text:
          description: >-
            The note/link text for `type == text`/`link`; for `type == image`
            or `file`, the optional caption the user added.
          type:
            - string
            - "null"
        download_url:
          description: >-
            Temporary, pre-signed download URL. Set only for `type == image`
            or `file`, and only valid for a limited time. For files:
            PDFs, plain text and JSON open in the browser, everything else
            is downloaded, either way under the original filename.
          type:
            - string
            - "null"
        file:
          description: Set only for `type == file`.
          oneOf:
            - $ref: "#/components/schemas/ListedFile"
            - type: "null"
        thumbnail_url:
          description: >-
            Temporary, pre-signed URL of a small WebP version of the image,
            meant for display in lists. Set only for `type == image`, once
            the thumbnail has been generated (shortly after upload); until
            then, or if generation failed, it is null and clients should
            fall back to `download_url`.
          type:
            - string
            - "null"

    ListItemsResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListedItem"
        next_cursor:
          description: Opaque cursor for the next page, or null when there is no more data.
          type:
            - string
            - "null"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer